♻️  Refactor AEMET module into separate stations and daily weather modules
Splits the monolithic aemet.py into focused modules for better maintainability. Updates CI workflow to use setup-uv@v6 and removes unnecessary parquet write options.
